Move graph iterator to seperate utility file
[ghstack-poisoned]
